backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
xgboost
lightgbm
chromadb
openai
langchain
langchain-openai
langchain-community
python-dotenv
plotly
matplotlib
seaborn
python-multipart
pydantic
httpx
scipy
openpyxl
